requirements.txt
asgiref==3.6.0
attr==0.3.2
attrs==22.2.0
blinker==1.6.2
Brotli==1.0.9
brotlicffi==1.0.9.2
configparser==5.3.0
cryptography==39.0.0
curio==1.6
Cython==0.29.33
dl==0.1.0
docutils==0.19
eventlet==0.33.3
flask_babel==3.1.0
flask_babelex==0.9.4
flask_uploads==0.2.1
gevent==22.10.2
greenlet==2.0.2
HTMLParser==0.0.2
httpx==0.24.0
importlib_metadata==4.12.0
inotify==0.2.10
ipython==8.13.2
ipywidgets==8.0.4
jnius==1.1.0
keyring==23.13.1
lockfile==0.12.2
mysql_connector_repackaged==0.3.1
numpy==1.23.3
ordereddict==1.1
pallets_sphinx_themes==2.1.0
PasteScript==3.3.0
Pillow==9.5.0
protobuf==4.23.0
pyOpenSSL==23.1.1
pytest==7.3.1
python-dotenv==1.0.0
railroad==0.5.0
requests_toolbelt==1.0.0
setproctitle==1.3.2
simplejson==3.18.4
sniffio==1.3.0
Sphinx==7.0.1
toml==0.10.2
tornado==6.2
trio==0.22.0
truststore==0.7.0
unicodedata2==15.0.0
urllib3_secure_extra==0.1.0
watchdog==3.0.0
wmi==1.5.1
xmlrpclib==1.0.1
zipp==3.8.1
